Native atom support for async defaults (#464)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/464

Add native atom support for async defaults.

Follow the [Async Atom Semantics](https://fb.quip.com/TUAhAfJI1CfJ).  Creating an atom with an async default set to a `Promise` will cause the atom to be in a "loading" state when read from components or selectors.  When the default resolves or rejects any subscribing components will re-render.  However, this will NOT be a state change and the resolved value will not be stored in state or cause observers to fire.

The `atomWithFallback` wrapper selector will only be used if the default is another `RecoilValue<T>`.  We may consider to remove that functionality from `atom()` and use an explicit helper utility instead.  However, we would want to consider this only after finishing migration of URL persistence to the new library that manages cross-atom backward compatibility with `restore` instead of fallback selectors.  We would also have to carefully consider composability issues, particularly for families and "get-on-set".

Reviewed By: davidmccabe

Differential Revision: D22367409

fbshipit-source-id: b7ad3e399528c78f3765b43e59fa5588661d9661